(IONotifyContoller): оптимизация функции localSetValue, попытка минимизировать, вызовы и блокировки.

+++ extensions/UNetUDP/unet-udp-tester.cc
@@ -302,7 +302,7 @@ int main(int argc, char* argv[])
 					mypack.addAData(d);
 				}
 
-				for( unsigned int i = 0; i < count; i++ )
+				for( size_t i = 0; i < count; i++ )
 					mypack.addDData(i, i);
 
 				Poco::Net::SocketAddress sa(s_host, port);
